Apply minor code improvements
