+++ app/build.gradle.kts
@@ -108,6 +108,7 @@ dependencies {
     implementation(libs.androidx.compose.ui.graphics)
     implementation(libs.androidx.compose.ui.text)
     implementation(libs.androidx.core.splashscreen)
+    implementation(libs.androidx.datastore.core)
     implementation(libs.androidx.datastore.preferences.core)
     implementation(libs.androidx.fragment)
     implementation(libs.androidx.lifecycle.common)
