+++ app/src/main/kotlin/com/oussamateyib/thoth/MainActivityState.kt
@@ -1,0 +1,20 @@
+package com.oussamateyib.thoth
+
+import com.oussamateyib.thoth.core.model.data.DarkThemeConfig
+import com.oussamateyib.thoth.core.model.data.UserData
+
+sealed interface MainActivityState {
+    data object Loading : MainActivityState
+    data class Success(
+        val userData: UserData
+    ) : MainActivityState {
+        override fun shouldUseDarkTheme(isSystemDarkTheme: Boolean) =
+            when (userData.darkThemeConfig) {
+                DarkThemeConfig.FOLLOW_SYSTEM -> isSystemDarkTheme
+                DarkThemeConfig.LIGHT -> false
+                DarkThemeConfig.DARK -> true
+            }
+    }
+
+    fun shouldUseDarkTheme(isSystemDarkTheme: Boolean) = isSystemDarkTheme
+}
