+++ app/src/main/kotlin/com/oussamateyib/thoth/MainActivityViewModel.kt
@@ -3,8 +3,6 @@ package com.oussamateyib.thoth
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.oussamateyib.thoth.core.domain.GetUserPreferencesStreamUseCase
-import com.oussamateyib.thoth.core.model.data.DarkThemeConfig
-import com.oussamateyib.thoth.core.model.data.UserData
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.map
@@ -16,27 +14,11 @@ class MainActivityViewModel @Inject constructor(
     getUserPreferencesStreamUseCase: GetUserPreferencesStreamUseCase
 ) : ViewModel() {
     val uiState = getUserPreferencesStreamUseCase().map {
-        MainActivityUiState.Success(it)
+        MainActivityState.Success(it)
     }.stateIn(
         scope = viewModelScope,
-        initialValue = MainActivityUiState.Loading,
+        initialValue = MainActivityState.Loading,
         // Wait 5 seconds before stopping to handle configuration changes smoothly
         started = SharingStarted.WhileSubscribed(5_000)
     )
-}
-
-sealed interface MainActivityUiState {
-    data object Loading : MainActivityUiState
-    data class Success(
-        val userData: UserData
-    ) : MainActivityUiState {
-        override fun shouldUseDarkTheme(isSystemDarkTheme: Boolean) =
-            when (userData.darkThemeConfig) {
-                DarkThemeConfig.FOLLOW_SYSTEM -> isSystemDarkTheme
-                DarkThemeConfig.LIGHT -> false
-                DarkThemeConfig.DARK -> true
-            }
-    }
-
-    fun shouldUseDarkTheme(isSystemDarkTheme: Boolean) = isSystemDarkTheme
 }
