+++ core/model/src/main/kotlin/com/oussamateyib/thoth/core/model/data/Language.kt
@@ -1,0 +1,13 @@
+package com.oussamateyib.thoth.core.model.data
+
+enum class Language(
+    val tag: String
+) {
+    FOLLOW_SYSTEM(""),
+    ENGLISH("en"),
+    ARABIC("ar");
+
+    companion object {
+        fun fromTag(tag: String) = entries.find { it.tag == tag } ?: FOLLOW_SYSTEM
+    }
+}
