+++ feature/settings/impl/src/main/kotlin/com/oussamateyib/thoth/feature/settings/impl/SettingsEvent.kt
@@ -1,7 +1,7 @@
 package com.oussamateyib.thoth.feature.settings.impl
 
 import com.oussamateyib.thoth.core.model.data.DarkThemeConfig
-import com.oussamateyib.thoth.core.ui.Language
+import com.oussamateyib.thoth.core.model.data.Language
 
 sealed class SettingsEvent {
     data class UpdateDarkThemeConfig(val config: DarkThemeConfig) : SettingsEvent()
